Make LogicalProcessor an inner class
