+++ oshi-core/src/main/java/oshi/hardware/CentralProcessor.java
@@ -330,4 +330,89 @@ public interface CentralProcessor extends Serializable {
             return index;
         }
     }
+
+    public class LogicalProcessor implements Serializable {
+
+        private static final long serialVersionUID = 1L;
+
+        private int processorNumber;
+        private int physicalProcessorNumber;
+        private int physicalPackageNumber;
+        private long currentFrequency;
+
+        /**
+         * The Logical Processor number as seen by the Operating System. Used
+         * for assigning process affinity and reporting CPU usage and other
+         * statistics.
+         * 
+         * @return the processorNumber
+         */
+        public int getProcessorNumber() {
+            return processorNumber;
+        }
+
+        /**
+         * @param processorNumber
+         *            the processorNumber to set
+         */
+        public void setProcessorNumber(int processorNumber) {
+            this.processorNumber = processorNumber;
+        }
+
+        /**
+         * The physical processor (core) id number assigned to this logical
+         * processor. Hyperthreaded logical processors which share the same
+         * physical processor will have the same number.
+         * 
+         * @return the physicalProcessorNumber
+         */
+        public int getPhysicalProcessorNumber() {
+            return physicalProcessorNumber;
+        }
+
+        /**
+         * @param physicalProcessorNumber
+         *            the physicalProcessorNumber to set
+         */
+        public void setPhysicalProcessorNumber(int physicalProcessorNumber) {
+            this.physicalProcessorNumber = physicalProcessorNumber;
+        }
+
+        /**
+         * The physical package (socket) id number assigned to this logical
+         * processor. Multicore CPU packages may have multiple physical
+         * processors which share the same number.
+         * 
+         * @return the physicalPackageNumber
+         */
+        public int getPhysicalPackageNumber() {
+            return physicalPackageNumber;
+        }
+
+        /**
+         * @param physicalPackageNumber
+         *            the physicalPackageNumber to set
+         */
+        public void setPhysicalPackageNumber(int physicalPackageNumber) {
+            this.physicalPackageNumber = physicalPackageNumber;
+        }
+
+        /**
+         * The current operating frequency of this logical processor, if known.
+         * Defaults to the vendor frequency.
+         * 
+         * @return the currentFrequency
+         */
+        public long getCurrentFrequency() {
+            return currentFrequency;
+        }
+
+        /**
+         * @param currentFrequency
+         *            the currentFrequency to set
+         */
+        public void setCurrentFrequency(long currentFrequency) {
+            this.currentFrequency = currentFrequency;
+        }
+    }
 }
