+++ oshi-core/src/main/java/oshi/hardware/common/AbstractCentralProcessor.java
@@ -31,7 +31,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import oshi.hardware.CentralProcessor;
-import oshi.hardware.LogicalProcessor;
 import oshi.util.ParseUtil;
 
 /**
